+++ .ebextensions/app.config
@@ -1,3 +1,3 @@
 option_settings:
   "aws:elasticbeanstalk:container:python":
-    WSGIPath: application.py
+    WSGIPath: manage.py
