+++ myapplication/database.py
@@ -1,8 +1,10 @@
 import time
 import logging
 from environs import Env
+from flask_sqlalchemy import SQLAlchemy
 import boto3
 from botocore.config import Config
+from myapplication import application
 
 env = Env()
 env.read_env()
@@ -94,3 +96,7 @@ class AWSPostgreSQL():
         uri += ':5432/'+env.str("AWS_DB")
 
         return uri
+
+application.config.from_object('config')
+db = SQLAlchemy(application)
+db.create_all()
