+++ myapplication/models.py
@@ -1,4 +1,5 @@
-from application import db
+from myapplication.database import db
+
 
 class Data(db.Model):
     id = db.Column(db.Integer, primary_key=True)
