+++ myapplication/templates/thanks.html
