+++ test.py
@@ -1,4 +1,4 @@
-from application import application
+from myapplication import application
 with application.test_client() as c:
     response = c.get('/')
     assert response.status_code == 200
